convert Vega component to functional component
